+++ src/index.js
@@ -7,28 +7,29 @@ import './js/refs/refs.js';
 import './js/layout/fetch-by-Id';
 import './js/layout/render-by-params';
 import './js/layout/fetch-by-keyword';
-import './js/layout/fetch-week-movies';
+import './js/layout/week-movies';
 import './js/layout/local-storage';
 import './js/layout/render-by-template';
 import './js/layout/render-storage-array';
 import './js/layout/switchTheme.js';
-import './js/layout/render-genres-list'
+import './js/layout/render-genres-list';
 
 //header
 import './js/layout/header';
+import './js/layout/filter';
 //main
+import './js/layout/pagination';
 
 //footer
 import './js/components/contacts';
 
 //=========components===============
 import './js/components/spiner.js';
-import './js/components/backdrop.js';
 import './js/components/modal.js';
 import './js/components/notifications.js';
 import './js/components/video-player.js';
 import './js/components/scroll-to-top';
-import './js/components/background'
+import './js/components/background';
 import './js/components/no-scroll.js';
 
 //=============api==================
@@ -39,4 +40,3 @@ import './js/test/test-api.js';
 import './js/test/test-video-player.js';
 import './js/test/test-modal.js';
 // import './js/test/test-notification.js';
-// import 'https://www.youtube.com/iframe_api';
