--- src/js/layout/fetch-week-movies.js
@@ -1,11 +1,0 @@
-// import imagesTpl from '../../partial/templates/film-cards.hbs'
-// import { refs } from '../refs/refs'
-// import {renderImages} from './render-images-to-main'
-
-// // --------- func for Main page and for Pagination -------------
-// export default function onFetchAllMovies(page) {
-//   let query = `/trending/movie/week?page=${page}`;
-//   renderImages(query, refs.filmsEl, imagesTpl )
-// }
-
-// onFetchAllMovies(1)
